+++ RestCrudOnServlet/src/main/java/com/campany/service/ManagerService.java
@@ -1,12 +1,16 @@
 package com.campany.service;
 
 import com.campany.dto.ManagerDTO;
+import com.campany.entity.Employee;
 import com.campany.entity.Manager;
+import com.campany.mapper.EmployeeMapper;
 import com.campany.mapper.ManagerMapper;
 import com.campany.repository.ManagerRepository;
+import lombok.extern.slf4j.Slf4j;
 
 import java.util.List;
 
+@Slf4j
 public class ManagerService implements CrudService<ManagerDTO> {
     private final ManagerRepository managerRepository;
     public ManagerService(ManagerRepository managerRepository) {
@@ -14,29 +18,66 @@ public class ManagerService implements CrudService<ManagerDTO> {
     }
     @Override
     public ManagerDTO getById(Integer id) {
-        Manager manager = managerRepository.findById(id);
-        return ManagerMapper.toDTO(manager);
+        log.info("Запрос информации о менеджере с id: {}.", id);
+        if (id != null) {
+            Manager manager = managerRepository.findById(id);
+            if (manager != null) {
+                return ManagerMapper.toDTO(manager);
+            } else {
+                log.warn("Менеджер с id: {} не найден.", id);
+                throw new RuntimeException("Сотрудник с id " + id + " не найден");
+            }
+        } else {
+            log.error("Неверно указан id для запроса информации о менеджере.");
+            throw new IllegalArgumentException("Неверно указан id для запроса информации о менеджере");
+        }
     }
     @Override
     public List<ManagerDTO> getAll() {
+        log.info("Запрос списка всех менеджеров.");
         List<Manager> managers = managerRepository.findAll();
         return ManagerMapper.toDTOList(managers);
     }
     @Override
-    public ManagerDTO create(ManagerDTO item) throws ClassNotFoundException {
-        Manager manager = ManagerMapper.toEntity(item);
-        managerRepository.save(manager);
-        return ManagerMapper.toDTO(manager);
+    public ManagerDTO create(ManagerDTO item) {
+        log.info("Создание нового менеджера.");
+        if (item != null) {
+            Manager manager = ManagerMapper.toEntity(item);
+            managerRepository.save(manager);
+            return ManagerMapper.toDTO(manager);
+        } else {
+            log.error("Невозможно создать менеджера: передан пустой объект ManagerDTO.");
+            throw new IllegalArgumentException("Невозможно создать менеджера: передан пустой объект ManagerDTO");
+        }
     }
     @Override
-    public ManagerDTO update(Integer id, ManagerDTO item) throws ClassNotFoundException {
-        item.setId(id);
-        Manager manager = ManagerMapper.toEntity(item);
-        managerRepository.update(manager);
-        return ManagerMapper.toDTO(manager);
+    public ManagerDTO update(Integer id, ManagerDTO item) {
+        log.info("Обновление информации о менеджере с id: {}.", id);
+        if (id != null && item != null) {
+            item.setId(id);
+            Manager manager = ManagerMapper.toEntity(item);
+            managerRepository.update(manager);
+            return ManagerMapper.toDTO(manager);
+        } else {
+            log.error("Невозможно обновить информацию о менеджере: указан неверный id или передан пустой объект ManagerDTO.");
+            throw new IllegalArgumentException("Невозможно обновить информацию о менеджере: указан неверный id или передан пустой объект ManagerDTO.");
+        }
     }
     @Override
-    public void deleteById(Integer id) throws ClassNotFoundException {
-        managerRepository.deleteById(id);
+    public void deleteById(Integer id) {
+        log.info("Удаление менеджера с id: {}.", id);
+        if (id != null) {
+            Manager manager = managerRepository.findById(id);
+            if (manager != null) {
+                managerRepository.deleteById(id);
+                log.info("Менеджер с id: {} удален успешно.", id);
+            } else {
+                log.warn("Менеджер с id: {} не найден, невозможно удалить.", id);
+                throw new RuntimeException("Менеджер с id " + id + " не найден, невозможно удалить");
+            }
+        } else {
+            log.error("Неверно указан id для удаления менеджера.");
+            throw new IllegalArgumentException("Неверно указан id для удаления менеджера");
+        }
     }
 }
